+++ api/konnect/v1alpha1/konnect_cloudgateway_network_types.go
@@ -123,8 +123,10 @@ type KonnectCloudGatewayNetworkStatus struct {
 }
 
 // GetKonnectAPIAuthConfigurationRef returns the Konnect API Auth Configuration Ref.
-func (c *KonnectCloudGatewayNetwork) GetKonnectAPIAuthConfigurationRef() konnectv1alpha2.KonnectAPIAuthConfigurationRef {
-	return c.Spec.KonnectConfiguration.APIAuthConfigurationRef
+func (c *KonnectCloudGatewayNetwork) GetKonnectAPIAuthConfigurationRef() konnectv1alpha2.ControlPlaneKonnectAPIAuthConfigurationRef {
+	return konnectv1alpha2.ControlPlaneKonnectAPIAuthConfigurationRef{
+		Name: c.Spec.KonnectConfiguration.APIAuthConfigurationRef.Name,
+	}
 }
 
 // KonnectCloudGatewayNetworkList contains a list of KonnectCloudGatewayNetwork.