+++ controller/konnect/watch_kongroute.go
@@ -65,7 +65,7 @@ func KongRouteReconciliationWatchOptions(
 			return b.Watches(
 				&configurationv1alpha1.KongReferenceGrant{},
 				handler.EnqueueRequestsFromMapFunc(
-					enqueueObjectForKongReferenceGrant[configurationv1alpha1.KongRouteList](cl),
+					enqueueObjectsForKongReferenceGrant[configurationv1alpha1.KongRouteList](cl),
 				),
 			)
 		},